♻️ Consolidate WithMinLen/MaxLen code
