+++ pkg/rules/rule_minlen_test.go
@@ -4,10 +4,12 @@ import (
 	"context"
 	"testing"
 
+	"proto.zip/studio/validate/pkg/errors"
 	"proto.zip/studio/validate/pkg/rules"
+	"proto.zip/studio/validate/pkg/testhelpers"
 )
 
-func TestMinLen(t *testing.T) {
+func TestSlice_MinLen(t *testing.T) {
 	ruleSet := rules.NewSlice[int]().WithMinLen(2)
 
 	// Prepare an output variable for Apply
@@ -38,7 +40,7 @@ func TestMinLen(t *testing.T) {
 // - Only one min length can exist on a rule set.
 // - Original rule set is not mutated.
 // - Most recent minimum is used.
-func TestMinLenConflict(t *testing.T) {
+func TestSlice_MinLen_Conflict(t *testing.T) {
 	ruleSet := rules.NewSlice[int]().WithMinLen(3).WithMaxLen(10)
 
 	// Prepare an output variable for Apply
@@ -66,13 +68,57 @@ func TestMinLenConflict(t *testing.T) {
 	}
 
 	// Verify that the original rule set's string representation is correct
-	expected := "ArrayRuleSet[int].WithMinLen(3).WithMaxLen(10)"
+	expected := "SliceRuleSet[int].WithMinLen(3).WithMaxLen(10)"
 	if s := ruleSet.String(); s != expected {
 		t.Errorf("Expected rule set to be %s, got %s", expected, s)
 	}
 
 	// Verify that the new rule set's string representation is correct
-	expected = "ArrayRuleSet[int].WithMaxLen(10).WithMinLen(2)"
+	expected = "SliceRuleSet[int].WithMaxLen(10).WithMinLen(2)"
+	if s := ruleSet2.String(); s != expected {
+		t.Errorf("Expected rule set to be %s, got %s", expected, s)
+	}
+}
+
+func TestString_WithMinLen(t *testing.T) {
+	ruleSet := rules.NewString().WithMinLen(2).Any()
+
+	testhelpers.MustApply(t, ruleSet, "abc")
+	testhelpers.MustApply(t, ruleSet, "ab")
+	testhelpers.MustNotApply(t, ruleSet, "a", errors.CodeMin)
+}
+
+// Requirements:
+// - Only one min length can exist on a rule set.
+// - Original rule set is not mutated.
+// - Most recent minimum is used.
+func TestString_WithMinLen_Conflict(t *testing.T) {
+	ruleSet := rules.NewString().WithMinLen(3).WithMaxLen(10)
+
+	// Prepare the output variable for Apply
+	var out string
+
+	// First validation with min length 3
+	if err := ruleSet.Apply(context.TODO(), "ab", &out); err == nil {
+		t.Errorf("Expected error to not be nil")
+	}
+	if err := ruleSet.Apply(context.TODO(), "abc", &out); err != nil {
+		t.Errorf("Expected error to be nil, got %s", err)
+	}
+
+	// Update the rule set with min length 2 and validate
+	ruleSet2 := ruleSet.WithMinLen(2)
+	if err := ruleSet2.Apply(context.TODO(), "ab", &out); err != nil {
+		t.Errorf("Expected error to be nil, got: %s", err)
+	}
+
+	// Check the string representation of the rule sets
+	expected := "StringRuleSet.WithMinLen(3).WithMaxLen(10)"
+	if s := ruleSet.String(); s != expected {
+		t.Errorf("Expected rule set to be %s, got %s", expected, s)
+	}
+
+	expected = "StringRuleSet.WithMaxLen(10).WithMinLen(2)"
 	if s := ruleSet2.String(); s != expected {
 		t.Errorf("Expected rule set to be %s, got %s", expected, s)
 	}
